requirements.txt
opencv-python
numpy
matplotlib
pillow
scikit-image
scikit-learn
streamlit
